requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
channels==4.3.1
channels_redis==4.3.0
Django==5.2.6
djangorestframework==3.16.1
msgpack==1.1.1
redis==6.4.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
